# api/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Gatte management API
  version: 1.6.0
  summary: The backend every Gatte front talks to (design/adr/0040).
  description: |
    Served by `mcp-gateway admin` over UNIX sockets only. There is no TCP
    listener. HTTP/1.1, JSON bodies, snake_case fields, times in RFC 3339
    UTC.

    Two sockets, two processes, two directories, two disjoint route sets.
    Every operation below carries `x-gatte-socket`:

    - `operator`: `/run/mcp-gateway-admin/operator/operator.sock`, served
      by a process running as the gateway's service account; the file is
      group `gatte-operators`, `0660`, in a directory of its own.
    - `accounts`: `/run/mcp-gateway-admin/accounts/accounts.sock`, served
      only by a root process; the file is `root:root 0600`, or
      `root:ACCOUNT_GROUP 0660` when `[admin] account_group` delegates
      accounts (both settings are required; the backend refuses a mismatch
      at startup). Its directory is `root:root 0755`.

    A route of the other socket answers 404 `wrong_socket` with
    `details.socket`; a route that does not exist answers 404
    `unknown_route`.

    ## Who serves

    `pkg/adminapi` reads the server's credentials after connecting and
    refuses an accounts socket whose server is not uid 0 (under systemd the
    kernel reports systemd, which created the socket). `WithServerUID` pins
    the operator socket's server.

    ## Who the operator is

...
    `GET /v1/whoami` returns what the backend decided.

...

...
servers:
...
components:
  parameters:
    GatteFront:
      name: Gatte-Front
      in: header
      required: false
      description: Label of the calling front, for the operator row's tag. Display only.
      schema:
        type: string
        pattern: '^[a-z0-9._-]{1,32}$'
...
  responses:
    Error:
      description: |
        Any failure. `code` is stable and machine-readable; `message` is
        one English sentence for a person and may change.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    ToolRef:
      type: object
      required: [server, tool]
      additionalProperties: false
      properties:
        server: {type: string}
        tool: {type: string}
...
    DeregisterRequest:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm: {type: string, description: The backend's name again, exactly. A mismatch is `invalid_argument` with `details.field` `confirm`.}
...
paths:
...
  /v1/overview:
    get:
      operationId: overview
      tags: [overview]
      x-gatte-socket: [operator]
      summary: What needs attention, the counts, and the latest refusals.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Overview'}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/tools/clear:
    post:
      operationId: clearTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: |
        Clear an approved sensitive tool for serving at its approved
        fingerprint (1.5.0, feature `tool_clear`, design/adr/0048). Writes
        `(tool clear)`.
      description: |
        The second decision a sensitive tool needs after approval: approval
        says the definition is not poisoned, clearing says this backend may
        act on the team's behalf through this operation. Refused unless some
        role marked `non_read` names the tool in `tools` (`no_non_read_grant`,
        with what covers it instead in `details.callable_by`): a wildcard or a
        `[role.grants]` name never reaches a sensitive tool, so a clearance
        nobody could use is not made. A change to the definition, or a
        revoke, withdraws the clearance with the approval. The gateway asks
        the same role question on every call.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ToolRef'}
      responses:
        '200':
          description: Cleared, or already cleared at exactly this fingerprint (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClearResult'}
        default:
          description: not_sensitive, not_approved, no_non_read_grant, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /v1/upstreams:
    get:
      operationId: listUpstreams
      tags: [upstreams]
      x-gatte-socket: [operator]
      summary: Registered backends, whether each is signed (`upstream list`) and, with `backend_health`, its health. Maintenance is under `/v1/maintenance`; with `console_manages`, register is `POST` here, deregister `DELETE /v1/upstreams/{name}` and sign is on the accounts socket.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [upstreams]
                properties:
                  upstreams:
                    type: array
                    items: {$ref: '#/components/schemas/Upstream'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: registerUpstream
      x-gatte-idempotent: false
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: Register a REST API from its OpenAPI document -- the same ingestion, refusals and report as `upstream register -transport http` (1.6.0, `console_manages`). Writes `(upstream register)`. Only http; stdio and oci stay in the terminal.
      description: |
        Not idempotent: a second call answers `upstream_exists`. After an
        `internal` or a dropped connection, `GET /v1/upstreams/{name}` first.
        The entry is NOT signed by this: sign it on the accounts socket.
        Each tool waits in the quarantine. Body at most 24 MiB.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterUpstreamRequest'}
      responses:
        '201':
          description: Registered; the ingestion's report.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegisterUpstreamResult'}
        default:
          description: feature_disabled, invalid_argument (any refusal of the ingestion, with its message), upstream_exists, payload_too_large, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/upstreams/{name}:
    get:
      operationId: getUpstream
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: One entry in detail -- for http its descriptor (names only) and each operation with method, path and class -- with its signature state and every tool the quarantine holds under the name (1.6.0, `console_manages`).
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UpstreamDetail'}
        default:
          description: feature_disabled, not_found, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      operationId: deregisterUpstream
      x-gatte-idempotent: true
      tags: [upstreams, manage]
      x-gatte-socket: [operator]
      summary: Deregister a backend, as `upstream deregister` does -- the registry row, its signature, every approval of its tools and its health, all keyed by the name (1.6.0, `console_manages`). Writes `(upstream deregister)`.
      description: |
        The body repeats the name (`confirm`); a front asks the operator to
        type it, because the approvals go with the entry. A name with
        nothing left under it answers `changed: false` and writes no row.
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DeregisterRequest'}
      responses:
        '200':
          description: Deregistered, or nothing was there.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeregisterResult'}
        default:
          description: feature_disabled, invalid_argument (confirm), internal (a later step failed; the earlier ones took effect and repeating is safe), ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/upstreams/{name}/sign:
    post:
      operationId: signUpstream
      x-gatte-idempotent: true
      tags: [upstreams, manage]
      x-gatte-socket: [accounts]
      summary: Sign an entry with the root signing key (`signer.key_file`), by a child `mcp-gateway sign` that reads the key as root and stores the signature as the database's owner (1.6.0, `console_manages`). Writes `(upstream sign)` with the key's fingerprint.
      description: |
        Every call signs again (a new row). The approval of each tool stays
        the human gate; a sensitive tool still needs `clearTool` and a role
        marked `non_read` naming it. `messages` carry what `sign` printed.
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: false
        content:
          application/json:
            schema: {type: object, additionalProperties: false}
      responses:
        '200':
          description: Signed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SignResult'}
        default:
          description: feature_disabled, not_found, invalid_argument (sign refused it, or no key_file), config_unavailable, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/secrets:
    get:
      operationId: listSecrets
      tags: [manage]
      x-gatte-socket: [accounts]
      summary: The vault's names -- never a value -- and, for each name a registered entry declares, whether the vault has it (1.6.0, `console_manages`).
      description: |
        The registry is read by a child running as the database's owner
        (root never opens the service account's database); when that
        fails, `registry_read` is false and only the vault's names are
        listed. The vault is decrypted by `sops` from standard input; its
        two files are held to the same ownership check as `config.toml`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SecretList'}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/serve-requests/{id}:
    get:
      operationId: serveRequest
      tags: [serve]
      x-gatte-socket: [operator]
      summary: Read a request to the gateway process back, e.g. one answered `pending`. Feature `serve_control`.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ServeRequest'}
        default: {$ref: '#/components/responses/Error'}
...